Change js in ts
