+++ package.json
@@ -4,6 +4,7 @@
   "private": true,
   "dependencies": {
     "@types/jest": "^24.0.21",
+    "@types/mapbox-gl": "^1.5.0",
     "@types/node": "^12.12.3",
     "@types/react": "^16.9.11",
     "@types/react-dom": "^16.9.3",
