--- src/components/Map/config.js
@@ -1,9 +1,0 @@
-export const token =
-    'pk.eyJ1Ijoib2xib2wiLCJhIjoiY2syZWhuNGoyMDF5MzNucGE5dDg4a2pmYyJ9.gOIAN2cwgDSCZ3GVgYnAAw';
-
-export const getInit = container => ({
-    container: container,
-    style: 'mapbox://styles/mapbox/streets-v9',
-    center: [30.318, 59.935],
-    zoom: 12,
-});
