+++ src/components/shared/Submit.tsx
@@ -11,6 +11,6 @@ const StyledSubmit = styled.button`
     background-color: #ffc617;
 `;
 
-const Submit = ({ name }) => <StyledSubmit type="submit">{name}</StyledSubmit>;
+const Submit = ({ name }: { name: string }) => <StyledSubmit type="submit">{name}</StyledSubmit>;
 
 export default Submit;
